implementando seeders, aplicando correção nos valores das tabelas (padronização) e testando controller

+++ database/migrations/2024_03_05_195426_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('price',6,2);
+            $table->decimal('price',6,2);
             $table->string('description','1000');
             $table->timestamps();
         });
